refactor: Components location
